# src/DocSwagger.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/OGLANEUX/portfolio/1.0.0
info:
  version: "1.0.0"
  title: portfolio Océane GLANEUX
  description: The API for the portfolio
  contact:
    name: Océane GLANEUX
    email: o.glaneux@gmail.com
paths:
...
  /Project:
    post:
      tags:
        - Project
      summary: Create a Project
      description: Create a new Project.
      operationId: CreateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
          application/xml:
            schema:
              $ref: '#/components/schemas/Project'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '201':
          description: Project Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
            application/xml:
              schema:
                $ref: '#/components/schemas/Project'
            application/x-www-form-urlencoded:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'          

    get:
      tags:
        - Project
      summary: Retrieve all Project
      description: Retrieve a list of all Project.
      operationId: getAllProject
      responses:
        '200':
          description: OK. The list of Project has been retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /certifications:
    get:
      tags:
        - Certification
      summary: Retrieve all certifications
      description: Retrieve a list of all certifications.
      operationId: getAllCertifications
      responses:
        '200':
          description: OK. The list of SkillCategories has been retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Certification'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Certification
      summary: Create a new certification
      description: Create a new certification.
      operationId: CreateCertification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Certification'   
          application/xml:
            schema:
              $ref: '#/components/schemas/Certification'   
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Certification'   
      responses:
        '201':
          description: Certification Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Certification'
            application/xml:
              schema:
                $ref: '#/components/schemas/Certification'
            application/x-www-form-urlencoded:
              schema:
                $ref: '#/components/schemas/Certification'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'   
  /certifications/{id}:
    put:
      tags:
        - Certification
      summary: Update a certification
      description: Update an existing certification.
      operationId: ModifyCertification
      parameters:
        - name: id
          in: path
          example: 12345    
          required: true
          description: ID of the Certification to retrieve
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Certification'
          application/xml:
            schema:
              $ref: '#/components/schemas/Certification'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Certification'
      responses:
        '201':
          description: SkillCategory Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Certification'
            application/xml:
              schema:
                $ref: '#/components/schemas/Certification'
            application/x-www-form-urlencoded:
              schema:
                $ref: '#/components/schemas/Certification'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'  

    delete:
      tags:
        - Certification
      summary: Delete a certification
      description: Delete an existing certification.
      operationId: deleteCertification
      parameters:
        - name: id
          in: path
          example: 12345    
          required: true
          description: ID of the Certification to retrieve
          schema:
            type: integer
            format: int64
      responses:
        '201':
          description: Certification deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags:
        - Certification
      summary: Retrieve a Certification
      description: Retrieve details about a specific Certification by its ID.
      operationId: getASCertification
      parameters:
        - name: id
          in: path
          example: 12345    
          required: true
          description: ID of the Certification to retrieve
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK. The Certification has been retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Certification'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  responses:
    NotFound:
      description: The specified resource was not found
...
    BadRequest:
      description: Bad request. The request could not be understood by the server due to malformed syntax, missing required parameters, or invalid data.
    Forbidden:
      description: The server understood the request, but refuses to authorize it. This may be due to insufficient permissions.
    InternalServerError:
      description: An unexpected error occurred on the server.
  schemas:
...
    Project:
      type: object
      required:
        - id
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          example: 1526
          description: "The unique identifier of the project."
        name:
          type: string
          example: "Project space"
          description: "The name of the project."
        shortDescription:
          type: string
          example: "Project space"
          description: "Additional details or description of the project."
        details:
          type: string
          description: "<html><body><ul><li>item 1</li><li>item 2</li></ul></body></html>"  
        links:
          nullable: true
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: "Project Link"
                description: "The name or title of the link."
              url:
                type: string
                format: uri
                example: "https://example.com"
                description: "The URL of the project link."
          example: 
            - name: "Project Link"
              url: "https://example.com"
          description: "List of links associated with the project."
        images:
          nullable: true
          type: array
          items:
            type: object
            properties:
              title:
                type: string
                example: "Project"
              alt:
                type: string
                example: "The code of my project"
              description:
                type: string
                example: "The code of my project"
              url:
                type: string
                example: "image.jpg"
          example: 
            - title: "Project"
              alt: "The code of my project"
              description: "The code of my project"
              url: "image.jpg"
          description: "List of images associated with the project."
        skills:
          type: array
          items:
            type: integer
            format: int64
            example: 12345
            description: "skill ID"
          uniqueItems: true
          example: [12345, 12346]
          description: "List of skill IDs associated with the education."
        created_at:
          type: string
          format: date-time
          example: "2022-01-15T13:45:30Z"
          description: "The date and time when the project was created."
        updated_at:
          type: string
          format: date-time
          example: "2022-03-20T10:30:15Z"
          description: "The date and time when the project was last updated."
...
    Certification:
      type: object
      required:
        - id
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          example: 1526
          description: "The unique identifier of the skill category."
        name:
          type: string
          example: "Certification Name"
          description: "The name of the certification."
        details:
          type: string
          example: "Certification Details"
          description: "Additional details or description of the certification."
        created_at:
          type: string
          format: date-time
          example: "2022-01-15T13:45:30Z"
          description: "The date and time when the skill category was created."
        updated_at:
          type: string
          format: date-time
          example: "2022-03-20T10:30:15Z"
          description: "The date and time when the skill category was last updated."
...